api/views/json: Move out of the "views" package
