+++ skylines/api/views/users.py
@@ -5,7 +5,7 @@ from werkzeug.exceptions import NotFound
 from skylines.api.schemas import user_schema, user_list_schema
 from skylines.api.views import pagination_args
 from skylines.model import User
-from skylines.api.views.json import jsonify
+from skylines.api.json import jsonify
 
 users = Blueprint('users', 'skylines')
